requirements.txt
# HTML/XML parsing
beautifulsoup4
lxml 

# dependency version for pytorch/tensorflow
protobuf==3.9.2
typing==3.6.6
grpcio==1.24.3
graphviz==0.8.1
urllib3==1.25.4

# bert-extractive-summarizer dependencies
tensorflow
torch==1.7.0
torchvision==0.8.1

spacy==2.1.3  # preprocessing
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz  # spacy model
neuralcoref
#transformers==2.2.2  # pre-trained lang models (e.g. BERT)
transformers==3.1.0
bert-extractive-summarizer

sentence-transformers  # sentence level embeddings
gensim  # topic modelling algs (e.g. LDA)
sklearn  # added for nearest neighbours

nbdime  # notebook merge tool
rouge-score  # evaluation metric
ipynb  # allows importing other notebooks
django==3.1.4
mysql-connector-python  # notebook mysql connector
mysqlclient  # django mysql connector
django-extensions  # for autogenerating slugs for django urls
django-autoslug
